Changed 2201 (Complement) to be prefix with high priority. It appears to be used in Set Theory as a rarely used form of A^c (https://en.wikipedia.org/wiki/Complement_(set_theory))
Changed 226c (between) to be the same precedence as the equality relations. This is a hard one to place. It is used in these papers
http://115.78.133.167:81/bitstream/TVHG_07113876976/1543/1/idoc.vn_approximation-for-the-preceding-and-crossing-structured-2-interval-pattern-problem.pdf
https://www.sciencedirect.com/science/article/pii/S0304397507004951

about "2-interval subsets". It is grouped with a version of subset and also < as an ordering relation on 2-intervals ('precedence order <', 'nest ⊏', or 'cross  ). So maybe it should be at 240 with set relations.

Alternatively, it is also used in an old book on "Kummer's Quartic Surface", but I don't see it in more modern descriptions of that. Someone says it denotes an inner product. Given its lack of modern usage in that context, I'm not placing it with inner product. You can see it's usage on page 28-29 in https://archive.org/details/quarticsurface00kummrich/page/n49/mode/2up

There were all mentioned in https://tex.stackexchange.com/questions/4771/what-is-the-symbol-between-%E2%89%AC-used-for