Remove decimal separator key symbol U+2396 from operator dictionary
From Wikipedia this does not seem to be an operator:

"This symbol is from ISO/IEC 9995 and is intended for use on a keyboard
to indicate a key that performs decimal separation."
(https://en.wikipedia.org/wiki/Decimal_separator#Other_numeral_systems)

Even if it is a decimal separator, such a separator is generally directly
used into the number together with the digits (e.g. <mn> in MathML) not
as an operator. Dictionary does not contain other similar separators like
the Arabic ones.

https://github.com/mathml-refresh/mathml/issues/176