Lots of changes: Sources: http://www.unicode.org/Public/math/latest/MathClassEx-15.html (referred to as Unicode or classex) https://gist.github.com/jaantollander/b49132a36b035cdfe05345dd0c7edf0f#file-unicode-math-symbols-csv-L1438 (referred to a TeX)
Removed 220e -- end of proof is not an operator
Removed 223f -- sine wave is not an operator
Double/Triple plus place with other '+'s
2206 -- increment -- place with unary plus (TeX has as mathord, Unicode unary)
27c2 -- perpendicular was missing (!) -- made same as 'parallel'

Removed 29B5 since seems to meant to be used only in superscript.
from https://chem.libretexts.org/Bookshelves/Physical_and_Theoretical_Chemistry_Textbook_Maps/DeVoes_Thermodynamics_and_Chemistry/07%3A_Pure_Substances_in_Single_Phases/77_Standard_States_of_Pure_Substances
"A superscript degree symbol (∘) denotes a standard quantity or standard-state conditions. An alternative symbol for this purpose, used extensively outside the U.S., is a superscript Plimsoll mark (⦵). (The Plimsoll mark is named after the British merchant Samuel Plimsoll, at whose instigation Parliament passed an act in 1875 requiring the symbol to be placed on the hulls of cargo ships to indicate the maximum depth for safe loading.)"

29B6/29b7 -- Only found one reference https://en.wiktionary.org/wiki/%E2%A6%B6/https://en.wiktionary.org/wiki/%E2%A6%B7 which says it is an electrical single/double socket symbol.
However, in keeping with the general principal that a circled version should be the same as the uncircled one and that the name is "CIRCLED PARALLEL", I've put 29b7 with parallel.
29b6 could go with either divides or vertical line.  Vertical line is tall, so I'll with grouping with divides.

29B8 -- found in https://www.cs.le.ac.uk/people/iulidowski/MGS/2014/RC2013.pdf which seems to place it higher than Union (p10). Placed with x29F5	reverse solidus operator in keeping with circled version goes with uncircled version which also meets the "above Union" constraint.

29b9 -- https://ipfs-sec.stackexchange.cloudflare-ipfs.com/tex/A/question/61871.html says they are sometimes used for orthogonal direct sums, which makes a bit of sense but I couldn't find actual usage. That would it at 280, but since I can't find usage, I'll put it with perpendicular at 260. This makes it a relation in contrast to Tex which makes it mathbin.

Removed 29c2, 29c3, 29c9, 29ba, 29bb, 29bd, 29be, 29bf, 29bd, 29ca, 29cb, 29cc, 29e0 as they don't appear to be operators (agrees with TeX). Nothing in arXiv to support them as operators.
Note: 29bc (circled anticlockwise-rotated division sign) is still in on the principal that it belongs with division since it is a circle version of it; TeX makes it 'mathord'.

Removed 29cd. Unicode points off to 25b3 for this symbol. Both of these are in TeX as mathbin, but no other upwards triangle is in MathML or TeX except for one that has multiplication in it. No references in arXiv and I can see no reason why TeX makes these mathbin instead of mathord like the other upwards triangles. Also Unicode classex doesn't list them as an operator.

Moved 29FB/29FC (double/triple plus) to be with pluses

The squared symbols 29c4, 29c5, 29c7, 29c8 are all infix (agrees with TeX) and are given precedence/spacing that correspondes with the inner symbol. For the slashes, this is division and sol

29ce, 29e1 moved to be relational op (agrees with TeX, Unicode)

29d6/29d7: TeX and MathClassEx.html both list these as binary operators. I can find no usages of them in searches. For lack of more info, I'll place them with times which is basically the next jump up from plus (which seems lower).

Moved bowties (22c8, 29d1, 29d2, 29d3) to relational ops group since both Tex and Unicode put them there

291D, 291E: both TeX and Unicode agree this should be a large op. Grouped with non sum/integral/product large ops.

2AF6: triple colon operator -- the unicode spec says it is used in logic, but I've been unable to find it anywhere, logic or outside of it despite it looking very familar. It could be used like a mapping, which would give it assymetrical spacing or used like ratio with symmetric spacing. TeX has it as mathbin, so I'll with symmetric spacing and a priority the same as 'ratio' although I'm unconvinced about the this.

2AFE: placed with infix vertical line (couldn't find any uses).

2999: It's name says "fence" and Unicode places it as a fence; TeX puts it as mathord. I'm going with fence.