Fixes to stretchy chars
Basic Philosophy:
The spec only talks about horizontal and vertical stretching, so no diagonal stretching. Those characters which have parts that can be repeated are marked as stretchy. Typically, that is a straight segment in the middle, but a wavy arrow can be repeated so those are also marked as stretchy.

The exception to the above rule are fence characters and large operators which might comne in a few predefined sizes, along with a few horizontally chars such as "hat".

For symmetric, only fences and large ops are marked as symmetric; arrow chars are not.

Removed stretchy from: 2196, 2197, 2198, 2199, 21D6, 21D7, 21D8, 21D9, 2921, 2922 (diagonal arrows), 21AF

Added stretchy to: 219A, 219B, 219E, 21CD, 21CE, 21CF, 21DE, 21DF, 21F7, 21F8, 21F9, 21FA, 21FB, 21FC (arrows with strokes),
    21F4, 2794, 2799, 279B - 27A1, 27A5, 27A6, 27A8 - 27B1,
    27B3, 27B5, 27B8, 27BA - 27BE, 27F4,
    2900 - 2909, 2911, 2914 - 2920,
    2934 - 2937 (I think they are stretchy in their initial part; I don't think they are corner arrows)
    2942 - 2948, 2971 - 2975, 2B40 - 2B44, 2B47 - 2B4C -- these are a little questionable, but I can easily imagine a way to stretch them although I doubt a font will do so
    2949, 294a - 294d, 2962 - 296df, 2970, 297C -297F, 2B04 - 2B07, 2B0C - 2B11, 2B30 - 2B3E
    2B45, 2B46, 2B60 - 2B65, 2B6A - 2B6D, 2B70 - 2B73, 2B7A - 2B7D, 2B80 - 2B87, 2B95, 2BA0 - 2BAF, 2BB8
    1F800 - 1F80B, 1F810 - 1F837, 1F844 - 1F853, 1F858 - 1F863, 1F868 - 1F86B, 1F870 - 1F873, 1F878 - 1F87B
    1F880 - 1F883, 1F898 - 1F8AB

Did not add stretchy to:
    2BB0 - 2BB7 (ribbon arrow down left, etc) as these looked like corner arrows
    1F838 - 1F83B (... squared arrow) -- seems like these were meant to be short
    1F83C - 1F843 (... compressed arrow) -- seems like these were meant to be short