Moved and/or blocks
Moved the "or"-like symbols to have the same priority as "plus"-like on the theory that and/or act like times/plus for predicate calculus. That's not true in other scenarios such as "a ^ b<c", but that seems less common.

I weaseled slight in making "and"-like symbols slightly lower than "times"-like so that having a bit more spacing makes sense.

Note: following Mathematica, "xor" is between the two groups.